NORMAL lots of changes:
Many bugfixes and modifications to the webbrowser and node support.

`require` and `node-require` now allow you to use the forms `define+`
and `define-macro+` to extend the global scope, which makes debugging
easier.

In general, the support for node.js is a lot better in this commit.

The biggest change is that Gazelle now supports object literal syntax
via a source pre-processing.  That is,

{ x 10 y 11 }

is now read as

({} x 10 y 11)

It should be safe to use this syntax even in macro expansion and
quoted code.

Finally the gazelle major mode has been extended to support much
better syntax highlighting and editing via paredit.